pyproject.toml
[project]
name = "trmnl-agenda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client<3",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "python-decouple==3.8",
    "requests<3",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 100
